backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
psycopg2-binary
pdfminer.six

# CPU-optimized ML dependencies (saves ~3GB vs GPU versions)
# Pin versions to prevent compatibility issues
numpy<2.0

# CPU-only PyTorch (much smaller than GPU version)
# Install index: https://download.pytorch.org/whl/cpu
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu
# torchvision removed - not needed for text embeddings
# torchaudio removed - not needed for text embeddings

transformers==4.40.1
sentence-transformers==2.7.0

google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
python-multipart
rank-bm25